# app/openapi_doc/faircombine.openapi.json
{"openapi":"3.0.2","info":{"title":"FAIR Combine API","description":"\nFAIR Combine is a web application designed to help users assess how FAIR their\ntheir Combine resources are.\n\nUsers may submit their Combine model, and the application will create a list\nof assessments following the FAIR principle. Some of these assessments will run\nin the background while others will need to be filled by the users.\n\nOnce all assessments are completed, the application returns a set of scores\ndescribing how FAIR their model is.\n","version":"0.0.1"},"paths":{"/session/create":{"get":{"tags":["Sessions"],"summary":"Create Session","operationId":"create_session_session_create_get","parameters":[{"required":true,"schema":{"title":"Subject Url","type":"string"},"name":"subject_url","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/session/load":{"post":{"tags":["Sessions"],"summary":"Load Session","operationId":"load_session_session_load_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/session/{session_id}":{"get":{"tags":["Sessions"],"summary":"Session Details","operationId":"session_details_session__session_id__get","parameters":[{"required":true,"schema":{"title":"Session Id","type":"string"},"name":"session_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/session/{session_id}/tasks/{task_id}":{"get":{"tags":["Assessments"],"summary":"Task Detail","operationId":"task_detail_session__session_id__tasks__task_id__get","parameters":[{"required":true,"schema":{"title":"Session Id","type":"string"},"name":"session_id","in":"path"},{"required":true,"schema":{"title":"Task Id","type":"string"},"name":"task_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/about_tasks":{"get":{"tags":["Assessments"],"summary":"Task Descriptions All","operationId":"task_descriptions_all_about_tasks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Task Descriptions All About Tasks Get","type":"array","items":{"$ref":"#/components/schemas/TaskDescription"}}}}}}}},"/about_tasks/{task_name}":{"get":{"tags":["Assessments"],"summary":"Task Description","operationId":"task_description_about_tasks__task_name__get","parameters":[{"required":true,"schema":{"title":"Task Name"},"name":"task_name","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskDescription"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/session/{session_id}/tasks/{task_id}/edit":{"post":{"tags":["Assessments"],"summary":"Update Task","operationId":"update_task_session__session_id__tasks__task_id__edit_post","parameters":[{"required":true,"schema":{"title":"Session Id","type":"string"},"name":"session_id","in":"path"},{"required":true,"schema":{"title":"Task Id","type":"string"},"name":"task_id","in":"path"},{"required":true,"schema":{"$ref":"#/components/schemas/TaskStatus"},"name":"task_status","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Session":{"title":"Session","required":["id","subject"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"subject":{"title":"Subject","anyOf":[{"maxLength":65536,"minLength":1,"type":"string","format":"uri"},{"type":"string","format":"file-path"}]},"tasks":{"title":"Tasks","type":"object","additionalProperties":{"$ref":"#/components/schemas/Task"},"default":{}},"status":{"allOf":[{"$ref":"#/components/schemas/SessionStatus"}],"default":"queued"},"score_essential":{"title":"Score Essential","type":"number"},"score_all":{"title":"Score All","type":"number"},"ratio_not_applicable":{"title":"Ratio Not Applicable","type":"number"}}},"SessionStatus":{"title":"SessionStatus","enum":["queued","preprocessing","running","postprocessing","finished","error"],"type":"string","description":"An enumeration."},"Task":{"title":"Task","required":["id","name","session_id"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"session_id":{"title":"Session Id","type":"string"},"children":{"title":"Children","type":"object","additionalProperties":{"$ref":"#/components/schemas/Task"},"default":{}},"priority":{"allOf":[{"$ref":"#/components/schemas/TaskPriority"}],"default":"essential"},"status":{"allOf":[{"$ref":"#/components/schemas/TaskStatus"}],"default":"queued"},"comment":{"title":"Comment","type":"string","default":""},"score":{"title":"Score","type":"number"}}},"TaskDescription":{"title":"TaskDescription","required":["name","priority","short","description"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"priority":{"title":"Priority","type":"string"},"short":{"title":"Short","type":"string"},"description":{"title":"Description","type":"string"}}},"TaskPriority":{"title":"TaskPriority","enum":["essential","important","useful"],"type":"string","description":"An enumeration."},"TaskStatus":{"title":"TaskStatus","enum":["queued","started","success","failed","warnings","error","not_applicable"],"type":"string","description":"An enumeration."},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}},"tags":[{"name":"Assessments","description":"FAIR Combine assessments. Endpoints allow to retrieve the details of a specific assessment associated with a session, to update an assessment status or to retrieve the documentation of FAIR Combine assessments."},{"name":"Sessions","description":"FAIR Combine assessment session. Endpoints to create a new session, to load a previously exported session, or to display the details of an existing session."}]}